requirements.txt
biopython==1.84
fastapi[standard]==0.135.3
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3
rdkit-pypi==2022.9.5
scikit-learn==1.5.2
